% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AminA.R
\name{AminA}
\alias{AminA}
\title{Translate any DNA or RNA sequence to IUPAC amino acid code}
\usage{
AminA(sequence)
}
\arguments{
\item{sequence}{a text string containing a DNA or RNA sequence}
}
\value{
A text string that represents the translated amino acid sequence
}
\description{
Translate any DNA or RNA sequence to IUPAC amino acid code
}
\examples{
sequence <- "ATGCTGTAG"
AminA(sequence)
}
